Gracefully skip files with errors; log errors to log file.
Also print some useful status messages to log file and terminal.

+++ build_database.py
@@ -10,6 +10,7 @@ import logging
 from configparser import ConfigParser
 from metadatasource import MongoDbDatabase
 from metadatafiletools import load_metadata
+from pluralize import pluralize
 
 
 ## INPUTS
@@ -50,7 +51,7 @@ class BeaverDB:
         logging.basicConfig(
             filename=directory / log_file_name,
             encoding="utf-8",
-            filemode='w',
+            filemode="w",
             format="%(levelname)s:%(message)s",
             level=logging.INFO,
         )
@@ -72,48 +73,70 @@ class BeaverDB:
 
     def update_database(self):
         """Add or update database information from each metadata file"""
+        # Print entry status
+        update_message = "{0} file{1} found in filesystem directory.".format(
+            len(self.input_files), pluralize(len(self.input_files))
+        )
+        print(update_message)
+        logging.info(update_message)
         # Store info to print exit status
         n_documents_deleted = 0
         n_new_documents = 0
+        n_skipped_files = 0
         # For each file in the list of files, manipulate it if needed then add it to the
         # database
         for input_file in tqdm(self.input_files):
-            # Load the file
+            # Load the file, if possible
             metadata_file = load_metadata(input_file)
-            # Get the ID for the file to use in the database
-            db_record_id = input_file.stem
+            if metadata_file:
+                # Get the ID for the file to use in the database
+                db_record_id = input_file.stem
 
-            # Convert the file to JSON if it isn't already
-            json_file = metadata_file.to_json()
-            # Add a meaningful _id tag for MongoDB
-            json_file["_id"] = db_record_id
+                # Convert the file to JSON if it isn't already
+                json_file = metadata_file.to_json()
+                # Add a meaningful _id tag for MongoDB
+                json_file["_id"] = db_record_id
 
-            # Update the database.  Note that some databases (e.g. MongoDB) have a
-            # single function which updates a record -or- creates a new record if the
-            # original one doesn't exist (in pymongo, this is update_one()).  However
-            # this doesn't work well when the _id field is included in the new document.
-            #
-            # Delete existing document from database if present
-            is_document_deleted = self.db.delete_single_record(db_record_id)
-            # Insert the new document
-            updated_document_id = self.db.insert_single_record(json_file)
-            # Record what happened
-            if is_document_deleted:
-                n_documents_deleted += 1
-            else:
-                n_new_documents += 1
+                # Update the database.  Note that some databases (e.g. MongoDB) have a
+                # single function which updates a record -or- creates a new record if
+                # the original one doesn't exist (in pymongo, this is update_one()).
+                # However this doesn't work well when the _id field is included in the
+                # new document.
+                #
+                # Delete existing document from database if present
+                is_document_deleted = self.db.delete_single_record(db_record_id)
+                # Insert the new document
+                updated_document_id = self.db.insert_single_record(json_file)
+                # Record what happened
+                if is_document_deleted:
+                    n_documents_deleted += 1
+                else:
+                    n_new_documents += 1
 
-            # Check that the updated document has the same _id as you intended
-            if updated_document_id != db_record_id:
-                logging.warning(
-                    "The document with _id = {0} was updated, instead of id = {1}.".format(
-                        updated_document_id, db_record_id
+                # Check that the updated document has the same _id as you intended
+                if updated_document_id != db_record_id:
+                    logging.warning(
+                        """The document with _id = {0} was updated, 
+                        instead of id = {1}.""".format(
+                            updated_document_id, db_record_id
+                        )
                     )
-                )
+            else:
+                logging.error("File {0} skipped due to file problems.".format(input_file))
+                n_skipped_files += 1
 
         # Report what happened
-        update_message = "Database updated!  {0} existing documents modified, {1} new documents added.".format(
-            n_documents_deleted, n_new_documents
+        update_message = (
+            "Database updated!"
+            + "\n{0} existing document{1} modified.".format(
+                n_documents_deleted, pluralize(n_documents_deleted)
+            )
+            + "\n{0} new document{1} added.".format(
+                n_new_documents, pluralize(n_new_documents)
+            )
+            + """\n{0} document{1} skipped due to file problems -- see beaverdam.log for details.""".format(
+                n_skipped_files, pluralize(n_skipped_files)
+            )
         )
         print(update_message)
         logging.info(update_message)
